backend/requirements.txt
asgiref==3.7.2
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
cryptography==42.0.5
defusedxml==0.8.0rc2
Django==4.2.7
django-cors-headers==4.3.1
django-extensions==3.2.3
django-oauth-toolkit==2.3.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
drf-social-oauth2==2.1.3
ecdsa==0.18.0
gunicorn==21.2.0
idna==3.6
jwcrypto==1.5.6
oauthlib==3.2.2
packaging==23.2
psycopg2-binary==2.9.9
pyasn1==0.5.1
pycparser==2.21
PyJWT==2.8.0
python-jose==3.3.0
python3-openid==3.2.0
pytz==2023.3.post1
requests==2.31.0
requests-oauthlib==1.4.0
rsa==4.9
six==1.16.0
social-auth-app-django==5.4.0
social-auth-core==4.5.3
sqlparse==0.4.4
typing_extensions==4.8.0
urllib3==2.2.1
